requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0

# LLM & Agent Orchestration
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langgraph==0.2.0

# Search & Research
tavily-python==0.4.0

# Vector Database (Memory)
chromadb==0.5.5

# Gmail API
google-api-python-client==2.140.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# CRM - Notion
notion-client==2.2.1

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# HTTP
httpx==0.27.0

# Templating
jinja2==3.1.4

# Utilities
python-multipart==0.0.9
